+++ src/doom/s_sound.cpp
@@ -265,8 +265,8 @@ void S_Init(int sfxVolume_param, int musicVolume_param)
     // (the maximum numer of sounds rendered
     // simultaneously) within zone memory.
     // [crispy] variable number of sound channels
-    channels = static_cast<decltype(channels)>(I_Realloc(nullptr, snd_channels * sizeof(channel_t)));
-    sobjs    = static_cast<decltype(sobjs)>(I_Realloc(nullptr, snd_channels * sizeof(degenmobj_t)));
+    channels = static_cast<decltype(channels)>(I_Realloc(nullptr, static_cast<unsigned long>(snd_channels) * sizeof(channel_t)));
+    sobjs    = static_cast<decltype(sobjs)>(I_Realloc(nullptr, static_cast<unsigned long>(snd_channels) * sizeof(degenmobj_t)));
 
     // Free all channels for use
     for (i = 0; i < snd_channels; i++)
@@ -434,7 +434,7 @@ void S_Start()
 
     // [crispy] do not change music if not changing map (preserves IDMUS choice)
     {
-        const short curmap = (gameepisode << 8) + gamemap;
+        const short curmap = static_cast<const short>((gameepisode << 8) + gamemap);
 
         if (prevmap == curmap || (nodrawers && singletics))
             return;
@@ -1090,8 +1090,8 @@ void S_UpdateSndChannels()
         snd_channels = 8;
     }
 
-    channels = static_cast<decltype(channels)>(I_Realloc(channels, snd_channels * sizeof(channel_t)));
-    sobjs    = static_cast<decltype(sobjs)>(I_Realloc(sobjs, snd_channels * sizeof(degenmobj_t)));
+    channels = static_cast<decltype(channels)>(I_Realloc(channels, static_cast<unsigned long>(snd_channels) * sizeof(channel_t)));
+    sobjs    = static_cast<decltype(sobjs)>(I_Realloc(sobjs, static_cast<unsigned long>(snd_channels) * sizeof(degenmobj_t)));
 
     for (i = 0; i < snd_channels; i++)
     {
